requirements.txt
ctransformers
sentence-transformers
langchain
langchain-community
langchain-core
langchain-text-splitters
pypdf
python-dotenv

flask

# --- Pinecone Stack ---
pinecone
langchain-pinecone